+++ DICOM/Network/DicomService.cs
@@ -1031,8 +1031,7 @@ namespace Dicom.Network
                             x.Result == DicomPresentationContextResult.Accept && x.AbstractSyntax == msg.SOPClassUID
                             && x.AcceptedTransferSyntax == dicomCStoreRequest.TransferSyntax);
                 if (pc == null)
-                    pc =
-                        Association.PresentationContexts.FirstOrDefault(
+                    pc = Association.PresentationContexts.FirstOrDefault(
                             x => x.Result == DicomPresentationContextResult.Accept && x.AbstractSyntax == msg.SOPClassUID);
             }
             else if (msg is DicomResponse)
@@ -1043,16 +1042,15 @@ namespace Dicom.Network
                 //fail safe if no presentation context is already assigned to the response (is this going to happen)
                 if (pc == null)
                 {
-                    pc =
-                        this.Association.PresentationContexts.FirstOrDefault<DicomPresentationContext>(
+                    pc = Association.PresentationContexts.FirstOrDefault(
                             x => x.Result == DicomPresentationContextResult.Accept && x.AbstractSyntax == msg.SOPClassUID);
                 }
             }
             else
             {
-                pc =
-                    Association.PresentationContexts.FirstOrDefault(
-                        x => x.Result == DicomPresentationContextResult.Accept && x.AbstractSyntax == msg.SOPClassUID);
+                var metaSopClasses = MetaSopClasses.GetMetaSopClass(msg.SOPClassUID);
+                pc = Association.PresentationContexts.FirstOrDefault(
+                        x => x.Result == DicomPresentationContextResult.Accept && metaSopClasses.Contains(x.AbstractSyntax));
             }
 
             if (pc == null)